+++ crates/sdk-core/tests/integ_tests/pagination_tests.rs
@@ -26,8 +26,7 @@ struct WeirdPaginationWf {
 impl WeirdPaginationWf {
     #[run(name = DEFAULT_WORKFLOW_TYPE)]
     async fn run(ctx: &mut WorkflowContext<Self>) -> WorkflowResult<()> {
-        ctx.wait_condition(|s| s.signal_count >= 2, Default::default())
-            .await?;
+        ctx.wait_condition(|s| s.signal_count >= 2).await?;
         Ok(())
     }
 
@@ -147,8 +146,7 @@ struct ExtremePaginationWf {
 impl ExtremePaginationWf {
     #[run(name = DEFAULT_WORKFLOW_TYPE)]
     async fn run(ctx: &mut WorkflowContext<Self>) -> WorkflowResult<()> {
-        ctx.wait_condition(|s| s.signal_count >= 6, Default::default())
-            .await?;
+        ctx.wait_condition(|s| s.signal_count >= 6).await?;
         Ok(())
     }